+++ src/jvmMain/kotlin/com/rsgkh/calendar/engine/ImmutableValue.kt
@@ -1,3 +1,4 @@
 package com.rsgkh.calendar.engine
 
 internal actual fun freezeValue(value: Any) = Unit
+internal actual fun requireInteger(value: Int) = Unit